Use ssh via vpn from metal-lib
